requirements.txt
mitreattack-python
bs4
